Conflicts:
	EatSenseRestWS/src/main/webapp/WEB-INF/appengine-web.xml

+++ EatSenseRestWS/src/main/java/net/eatsense/auth/AccessTokenFilter.java
@@ -15,6 +15,7 @@ import net.eatsense.domain.Account;
 import net.eatsense.exceptions.IllegalAccessException;
 import net.eatsense.exceptions.NotFoundException;
 import net.eatsense.persistence.AccountRepository;
+import net.eatsense.persistence.CheckInRepository;
 
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
@@ -30,6 +31,8 @@ import com.sun.jersey.spi.container.ContainerRequestFilter;
  *
  */
 public class AccessTokenFilter implements ContainerRequestFilter {
+	private static final String BUSINESS_PATH_PREFIX = "b";
+
 	protected Logger logger = LoggerFactory.getLogger(this.getClass());
 	
 	public final static String TOKEN_HEADER = "X-Auth";
@@ -43,13 +46,16 @@ public class AccessTokenFilter implements ContainerRequestFilter {
     HttpServletRequest servletRequest;
 
 	private final AuthorizerFactory authorizerFactory;
+
+	private final CheckInRepository checkInRepo;
 	
 	@Inject
-	public AccessTokenFilter(AccessTokenRepository accessTokenRepo, AccountRepository accountRepo, AuthorizerFactory authorizerFactory) {
+	public AccessTokenFilter(AccessTokenRepository accessTokenRepo, AccountRepository accountRepo, AuthorizerFactory authorizerFactory, CheckInRepository checkInRepo) {
 		super();
 		this.authorizerFactory = authorizerFactory;
 		this.accountRepo = accountRepo;
 		this.accessTokenRepo = accessTokenRepo;
+		this.checkInRepo = checkInRepo;
 	}
 
 	@Override
@@ -57,28 +63,6 @@ public class AccessTokenFilter implements ContainerRequestFilter {
 		String stringToken = request.getHeaderValue(TOKEN_HEADER);
 		//AbstractResourceMethod method = resourceContext.matchUriInfo(request.getRequestUri()).getMatchedMethod();
 		//TokenType requiredToken = null;
-		
-		Long businessId = null;
-		
-		for (Iterator<PathSegment> iterator = request.getPathSegments(true).iterator(); iterator.hasNext();) {
-			PathSegment pathSegment = iterator.next();
-			if(pathSegment.getPath().equals("businesses") ) {
-				try {
-					if(iterator.hasNext())
-						businessId = Long.valueOf(iterator.next().getPath());
-				} catch (NumberFormatException e) {
-				}
-			}
-		}
-		
-		if(businessId == null) {
-			if( request.getFormParameters().getFirst("businessId") != null)
-				try {
-					businessId = Long.valueOf(request.getFormParameters().getFirst("businessId"));
-				} catch (NumberFormatException e) {
-					logger.error("businessId invalid");
-				}
-		}
 		
 //		if(method != null) {
 //			TokenRequired tr = method.getAnnotation(TokenRequired.class);
@@ -95,11 +79,18 @@ public class AccessTokenFilter implements ContainerRequestFilter {
 				throw new IllegalAccessException("Access token invalid, please re-authenticate.");
 			}
 			
-			if(accessToken.getExpires().before(new Date()) ) {
+			if( accessToken.getExpires() != null &&  accessToken.getExpires().before(new Date()) ) {
 				accessTokenRepo.delete(accessToken);
 				logger.info("Deleted expired access token for account {}", accessToken.getAccount());
 				throw new IllegalAccessException("Access token exired, please re-authenticate.");	
 			}
+			
+			if( accessToken.getType() == TokenType.AUTHENTICATION_CUSTOMER) {
+				if(request.getPathSegments().get(0).equals(BUSINESS_PATH_PREFIX)) {
+					throw new IllegalAccessException("Can not access business resource with this access token.");
+				}
+			}
+			
 			
 			Account account = accountRepo.getByKey(accessToken.getAccount());
 			
@@ -121,9 +112,18 @@ public class AccessTokenFilter implements ContainerRequestFilter {
 //
 			servletRequest.setAttribute(AccessToken.class.getName(), accessToken);
 			
-			if(accessToken.getType() == TokenType.AUTHENTICATION) {
+			if(accessToken.getType() == TokenType.AUTHENTICATION || accessToken.getType() == TokenType.AUTHENTICATION_CUSTOMER) {
 				request.setSecurityContext(authorizerFactory.createForAccount(account, accessToken));
 				servletRequest.setAttribute("net.eatsense.domain.Account", account);
+				if(account.getActiveCheckIn() != null) {
+					try {
+						servletRequest.setAttribute("net.eatsense.domain.CheckIn", checkInRepo.getByKey(account.getActiveCheckIn()));
+					} catch (com.googlecode.objectify.NotFoundException e) {
+						logger.info("activeCheckin for account not found, removing reference");
+						account.setActiveCheckIn(null);
+						accountRepo.saveOrUpdate(account);
+					}
+				}
 				logger.info("Request authenticated success for user: "+account.getLogin());
 			}
 		}
